Arduino sketch that returns calibration offsets for MPU6050
These offsets were meant to calibrate MPU6050's internal DMP, but can be also useful for reading sensors. 
The effect of temperature has not been taken into account so I can't promise that it will work if you 
calibrate indoors and then use it outdoors. Best is to calibrate and use at the same room temperature.